+++ app/src/main/java/org/breezyweather/ui/details/components/DetailsCommon.kt
@@ -23,13 +23,19 @@ import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.wrapContentHeight
+import androidx.compose.foundation.text.InlineTextContent
+import androidx.compose.foundation.text.TextAutoSize
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.automirrored.outlined.HelpOutline
 import androidx.compose.material3.HorizontalDivider
 import androidx.compose.material3.Icon
+import androidx.compose.material3.LocalTextStyle
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.PlainTooltip
 import androidx.compose.material3.Text
@@ -40,13 +46,24 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.clearAndSetSemantics
+import androidx.compose.ui.text.AnnotatedString
+import androidx.compose.ui.text.TextLayoutResult
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
 import kotlinx.coroutines.launch
 import org.breezyweather.R
@@ -147,10 +164,20 @@ fun UnavailableChart(
 }
 
 @Composable
-fun NighttimeWithInfo(
+fun DaytimeLabel(
     modifier: Modifier = Modifier,
 ) {
-    val context = LocalContext.current
+    TextFixedHeight(
+        text = stringResource(R.string.daytime),
+        style = MaterialTheme.typography.labelMedium,
+        modifier = modifier
+    )
+}
+
+@Composable
+fun NighttimeLabelWithInfo(
+    modifier: Modifier = Modifier,
+) {
     val tooltipState = rememberTooltipState()
     val coroutineScope = rememberCoroutineScope()
 
@@ -167,25 +194,25 @@ fun NighttimeWithInfo(
                     coroutineScope.launch {
                         tooltipState.show()
                     }
-                },
+                }
+                .height(
+                    with(LocalDensity.current) {
+                        MaterialTheme.typography.labelMedium.lineHeight.toDp()
+                    }
+                ),
             horizontalArrangement = Arrangement.spacedBy(dimensionResource(R.dimen.little_margin)),
             verticalAlignment = Alignment.CenterVertically
         ) {
             Text(
                 text = stringResource(R.string.nighttime),
-                style = MaterialTheme.typography.labelMedium
+                style = MaterialTheme.typography.labelMedium,
+                maxLines = 1
             )
             Icon(
                 imageVector = Icons.AutoMirrored.Outlined.HelpOutline,
                 contentDescription = null, // TooltipBox already takes care of adding the info that there is a tooltip
                 modifier = Modifier
-                    .size(
-                        context.pxToDp(
-                            context.spToPx(
-                                MaterialTheme.typography.labelMedium.fontSize.value.roundToInt()
-                            ).roundToInt()
-                        ).dp
-                    )
+                    .fillMaxHeight()
             )
         }
     }
@@ -255,4 +282,104 @@ fun DetailsCardText(
             }
         }
     }
+}
+
+@Composable
+fun TextFixedHeight(
+    text: String,
+    modifier: Modifier = Modifier,
+    color: Color = Color.Unspecified,
+    autoSize: TextAutoSize? = null,
+    fontSize: TextUnit = TextUnit.Unspecified,
+    fontStyle: FontStyle? = null,
+    fontWeight: FontWeight? = null,
+    fontFamily: FontFamily? = null,
+    letterSpacing: TextUnit = TextUnit.Unspecified,
+    textDecoration: TextDecoration? = null,
+    textAlign: TextAlign? = null,
+    lineHeight: TextUnit = TextUnit.Unspecified,
+    overflow: TextOverflow = TextOverflow.Clip,
+    softWrap: Boolean = true,
+    maxLines: Int = 1,
+    minLines: Int = 1,
+    onTextLayout: ((TextLayoutResult) -> Unit)? = null,
+    style: TextStyle = LocalTextStyle.current,
+) {
+    Text(
+        text,
+        modifier
+            .height(
+                with(LocalDensity.current) {
+                    style.lineHeight.times(maxLines).toDp()
+                }
+            )
+            .wrapContentHeight(align = Alignment.CenterVertically),
+        color,
+        autoSize,
+        fontSize,
+        fontStyle,
+        fontWeight,
+        fontFamily,
+        letterSpacing,
+        textDecoration,
+        textAlign,
+        lineHeight,
+        overflow,
+        softWrap,
+        maxLines,
+        minLines,
+        onTextLayout,
+        style
+    )
+}
+
+@Composable
+fun TextFixedHeight(
+    text: AnnotatedString,
+    modifier: Modifier = Modifier,
+    color: Color = Color.Unspecified,
+    autoSize: TextAutoSize? = null,
+    fontSize: TextUnit = TextUnit.Unspecified,
+    fontStyle: FontStyle? = null,
+    fontWeight: FontWeight? = null,
+    fontFamily: FontFamily? = null,
+    letterSpacing: TextUnit = TextUnit.Unspecified,
+    textDecoration: TextDecoration? = null,
+    textAlign: TextAlign? = null,
+    lineHeight: TextUnit = TextUnit.Unspecified,
+    overflow: TextOverflow = TextOverflow.Clip,
+    softWrap: Boolean = true,
+    maxLines: Int = 1,
+    minLines: Int = 1,
+    inlineContent: Map<String, InlineTextContent> = mapOf(),
+    onTextLayout: (TextLayoutResult) -> Unit = {},
+    style: TextStyle = LocalTextStyle.current,
+) {
+    Text(
+        text,
+        modifier
+            .height(
+                with(LocalDensity.current) {
+                    style.lineHeight.times(maxLines).toDp()
+                }
+            )
+            .wrapContentHeight(align = Alignment.CenterVertically),
+        color,
+        autoSize,
+        fontSize,
+        fontStyle,
+        fontWeight,
+        fontFamily,
+        letterSpacing,
+        textDecoration,
+        textAlign,
+        lineHeight,
+        overflow,
+        softWrap,
+        maxLines,
+        minLines,
+        inlineContent,
+        onTextLayout,
+        style
+    )
 }