+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 **Improvements and fixes**
 - Details page - Fix floating action button not updating in real time (@min7-i)
 - Details page - Charts - Fix area fill in Right to Left languages (@chunshek)
+- Details page - Fix jumping of the chart when tapping on it
 - Fix tint of “Open in another app” icon in landscape mode
 - Improve the formatting of today/tomorrow notification
 - Conditions screen: move long weather condition description to a dedicated Daily summary card (especially noticeable with AccuWeather source)